requirements.txt
openai==1.30.5
pdfplumber==0.11.4
PyMuPDF==1.24.10
pydantic==2.7.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
requests==2.31.0
httpx==0.27.0
rich==13.7.1
tenacity==8.3.0
